refactor(panel): keep RelationFieldInput atom router-free
Per .claude/instructions/atomic-design.md, atoms must be stateless and free of
side effects / business logic. RelationFieldInput (an atom) had imported
react-router-dom and implemented relation-stack navigation directly. Move that
logic up into the BucketEntryDrawer organism, which builds an onNavigate(documentId)
callback (target bucket + history-state Back trail) and passes it down; the atom
now just renders the "open" chips and invokes the plain onNavigate prop.
